pyproject.toml
[tool.poetry]
name = "result-reporting"
version = "0.1.0"
description = "Election Reporting for VA08"
authors = ["Bentley Hensel <bentleyhensel@gmail.com>"]
license = "AGPL-3.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
flask = "^3.0.3"
requests = "^2.32.3"
apscheduler = "^3.10.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
